backend/requirements.txt
fastapi[standard]
httpx
pydantic-settings
pytest
python-ulid[pydantic]
sqlmodel